refactor: server - extracting module file
will be used to interface with prediction-model; separate module will make testing easier
